==> code/5-tree-modeling.Rmd <==
---
title: "R Notebook"
output:
  pdf_document: default
  html_notebook: default
---

```{r}
# load libraries
library(kableExtra)                     # for printing tables
library(cowplot)                        # for side by side plots
library(glmnetUtils)                    # to run ridge and lasso
library(tidyverse)                      # for everything else
library(naniar)                         # for replace NA 
library(rpart)
library(rpart.plot)
library(randomForest)
library(gbm)
library(tidyverse)
library(nnet)
library(pROC)
source("../code/functions/plot_glmnet.R")   
```

```{r}
# read in the training data
vh_train = read_tsv("../data/clean/vh_train.tsv")

# convert specific features to factor type
cols = c("region", "hispanic", "race", "education_level", "marital_status", 
         "birth_gender", "described_gender", "sexual_orientation", "had_covid", 
         "recent_work_loss", "employment", "child_tax_credit", 
         "difficulty_with_expenses", "work_onsite", "work_from_home", 
         "in_store_shopping", "eating_indoors_restaurants", 
         "in_person_medical_appointments", "in_home_housekeeping", 
         "household_food_sufficiency", "household_free_food", "snap_receipt",
         "anxious", "worry", "interest", "depressed", 
         "prescription_mental_health", "mental_health_services", 
         "mental_health_not_get", "private_health_insurance", 
         "public_health_insurance", "telehealth_personal", "telehealth_child",
         "child_missed_checkups", "seeing_limitations", "hearing_limitations", 
         "remembering_limitations", "mobility_limitations", 
         "housing_owned_rented", "building_type", "rent_caught_up", 
         "mortgage_caught_up", "housing_pay_confidence", 
         "rent_assistance_apply", "eviction_likelihood", "foreclose_likelihood",
         "energy_pay_trouble", "housing_unsafe_temp", "income", "kids", 
         "active_duty", "essential_worker", "pandemic_child_impact", 
         "summer_educational_catch_up", "post_sec_edu_plan_changes", 
         "children_public_school", "children_private_school", 
         "children_homeschool", "post_secondary_classes"
         )
vh_train <-
  vh_train %>%
       mutate_each_(funs(factor(.)), cols)
```

# Classification tree

```{r}
# run classification tree 
deepest_classification_tree_fit = rpart(vaccine_hesitancy ~ .,
                 method = "class", # classification
                 parms = list(split = "gini"), # Gini index for splitting 
                 control = rpart.control(minsplit = 1, minbucket = 1, cp = 0),
                 data = vh_train)

# produce CV plot
cp_table = printcp(deepest_decision_tree_fit) %>% as_tibble()
cp_table %>%
  ggplot(aes(x = nsplit+1, y = xerror,
             ymin = xerror - xstd, ymax = xerror + xstd)) +
  geom_point() + geom_line() +
  geom_errorbar(width = 0.2) +
  xlab("Number of terminal nodes") + ylab("CV error") +
  geom_hline(aes(yintercept = min(xerror)), linetype = "dashed") +
  theme_bw()

# construct optimal classification tree
optimal_tree_info = cp_table %>%
  filter(xerror - xstd < min(xerror)) %>%
  arrange(nsplit) %>%
  head(1)
optimal_tree_info

# prune optimal tree fit
optimal_tree_fit = prune(tree = decision_tree_fit, cp = optimal_tree_info$CP)

# save optimal tree fit
save(optimal_tree_fit, file = "../results/optimal_tree_fit.Rda")

# save optimal tree plot
png(width = 6, 
    height = 4,
    res = 300,
    units = "in", 
    filename = "../results/classification-tree-plot.png")
rpart.plot(optimal_tree_fit)
dev.off()
```

# Boosted model

```{r}
# train boosted model
set.seed(1)
boosted_fit = gbm(vaccine_hesitancy ~ .,
              distribution = "bernoulli",
              n.trees = 500,
              interaction.depth = 1,
              shrinkage = 0.1,
              cv.folds = 5,
              data = vh_train)
opt_num_trees = gbm.perf(boosted_fit)

# test values for interaction depth
set.seed(1)
boosted_fit_1 = gbm(vaccine_hesitancy ~ .,
              distribution = "bernoulli",
              n.trees = 500,
              interaction.depth = 1,
              shrinkage = 0.1,
              cv.folds = 5,
              data = vh_train)
set.seed(1)
boosted_fit_2 = gbm(vaccine_hesitancy ~ .,
              distribution = "bernoulli",
              n.trees = 500,
              interaction.depth = 2,
              shrinkage = 0.1,
              cv.folds = 5,
              data = vh_train)
set.seed(1)
boosted_fit_3 = gbm(vaccine_hesitancy ~ .,
              distribution = "bernoulli",
              n.trees = 500,
              interaction.depth = 3,
              shrinkage = 0.1,
              cv.folds = 5,
              data = vh_train)
set.seed(1)
boosted_fit_4 = gbm(vaccine_hesitancy ~ .,
              distribution = "bernoulli",
              n.trees = 500,
              interaction.depth = 4,
              shrinkage = 0.1,
              cv.folds = 5,
              data = vh_train)

# extract CV errors
ntrees = 500
cv_errors = bind_rows(
  tibble(ntree = 1:ntrees, cv_err = boosted_fit_1$cv.error, depth = 1),
  tibble(ntree = 1:ntrees, cv_err = boosted_fit_2$cv.error, depth = 2),
  tibble(ntree = 1:ntrees, cv_err = boosted_fit_3$cv.error, depth = 3),
  tibble(ntree = 1:ntrees, cv_err = boosted_fit_4$cv.error, depth = 4)
)

# plot CV errors
cv_error_plot <-
  cv_errors %>%
  ggplot(aes(x = ntree, y = cv_err, colour = factor(depth))) +
  geom_line() + 
  theme_bw() +
  ggtitle("Tuning Interaction Depth") + 
  theme(plot.title = element_text(hjust = 0.5))

# save CV error plot
ggsave(filename = "../results/boosted-cv-plot.png", 
       plot = cv_error_plot, 
       device = "png", 
       width = 6, 
       height = 4)

# identify optimal interaction depth and number of trees
boosted_fit_optimal = boosted_fit_3
optimal_num_trees = gbm.perf(boosted_fit_3, plot.it = FALSE)

# save optimal boosted fit
save(boosted_fit_optimal, file = "../results/optimal_boosted_fit.Rda")

# save summary of boosted variable importance
boosted_var_summary <-
  summary(boosted_fit_optimal, n.trees = optimal_num_trees, plotit = FALSE) %>%
  head(10) %>%
  write_tsv("../results/boosted-var-summary.tsv")

# save age partial dependence plot 
png(width = 6, 
    height = 4,
    res = 300,
    units = "in", 
    filename = "../results/boosted-age-var-plot.png")
plot(boosted_fit_optimal, i.var = "age", n.trees = optimal_num_trees)
dev.off()

# save education partial dependence plot 
png(width = 6, 
    height = 4,
    res = 300,
    units = "in", 
    filename = "../results/boosted-education-var-plot.png")
plot(boosted_fit_optimal, i.var = "education_level", 
     n.trees = optimal_num_trees)
dev.off()

# save covid partial dependence plot 
png(width = 6, 
    height = 4,
    res = 300,
    units = "in", 
    filename = "../results/boosted-covid-var-plot.png")
plot(boosted_fit_optimal, i.var = "had_covid", n.trees = optimal_num_trees)
dev.off()
```
